moved provider specific validations into interface
